refactor: rename settings to options
